Vanished players can no longer block merchant trading
The trading system in vanilla is implemented with just one trading
player in mind. Because of that the trade screen of vanished players is
simply closed when another player attempts to trade with the villager to
make it seem like nobody else is trading with the villager to non
vanished players. These changes have low priority and may be removed at
any time, because it is easy to not allow this case to happen as a
vanished player.
resolves #46